Add files via upload
This website was made by Hugo Burton. I recreated the Tesla home page with my own twist to it. This homepage was made using only HTML and CSS , because of this the effect thats is on the nav a on the Original Tesla homepage is impossible to make without the knowledge of advanced JAVASCRIPT. I hope you like the Tesla home page i made :)